requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
plotly>=5.18.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
alembic>=1.12.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
httpx>=0.24.0
watchdog>=3.0.0
tenacity>=8.2.0
rich>=13.0.0
PyJWT>=2.8.0
typer>=0.9.0